move getLocalTime to time.cpp
