--- cores/esp8266/timehelper.cpp
@@ -1,37 +1,0 @@
-/*
- * timehelper.c - generic time functions
- * Copyright (c) 2021 esp8266/Arduino. All rights reserved.
- * This file is part of the esp8266 core for Arduino environment.
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details.
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
- *
- */
-
-
-#include "Arduino.h"
-
-// https://github.com/espressif/arduino-esp32/blob/master/cores/esp32/esp32-hal-time.c
-
-bool getLocalTime(struct tm * info, uint32_t ms)
-{
-    uint32_t start = millis();
-    time_t now;
-    while((millis()-start) <= ms) {
-        time(&now);
-        localtime_r(&now, info);
-        if(info->tm_year > (2016 - 1900)){
-            return true;
-        }
-        delay(10);
-    }
-    return false;
-}
